edited ccomments in new group model, moved folder inside components directory and updated impacted imports
