+++ client/src/components/NewGroupModal/NewGroupModal.js
@@ -3,9 +3,9 @@ import Popup from "reactjs-popup";
 import "reactjs-popup/dist/index.css";
 import "./NewGroupModal.css";
 import { Avatar } from "@material-ui/core";
-import { UserContext } from "../utils/UserStore";
-import { GroupContext } from "../utils/GroupStore";
-import API from "../utils/API";
+import { UserContext } from "../../utils/UserStore";
+import { GroupContext } from "../../utils/GroupStore";
+import API from "../../utils/API";
 import GroupAddIcon from "@material-ui/icons/GroupAdd";
 
 function NewGroupModal() {
@@ -14,14 +14,7 @@ function NewGroupModal() {
   const { userState, setUserState } = useContext(UserContext);
   const { groupState, setGroupState } = useContext(GroupContext);
 
-  // console.log(groupState.groupMembers)
-  //console.log("after rata");
-  // access array, .map
   let subGroupData = {};
-
-  //   useEffect(() => {
-  //       loadSubGroup(users)
-  //   }, [])
 
   function refreshUser(username) {
     // console.log("before");
@@ -34,6 +27,7 @@ function NewGroupModal() {
       .catch((err) => console.log(err));
   }
 
+  // Builds array of objects with all info for Users in new Subgroup
   function buildUserForm(event) {
     console.log(event.target);
     const id = event.target.value;
@@ -54,6 +48,7 @@ function NewGroupModal() {
     console.log(input);
   }
 
+  // Creates new subgroup then calls functions to add appropriate data to other group and user documents
   function handleFormSubmit(e) {
     e.preventDefault();
     console.log(newSubGroup);
@@ -77,6 +72,7 @@ function NewGroupModal() {
       .catch((err) => console.log(err));
   }
 
+  // Adds new subgroup info to User documents
   function addSubGroupToUsers(users, subGroupData) {
     let totalCount = users.length;
     let completedCount = 0;
@@ -92,17 +88,8 @@ function NewGroupModal() {
         .catch((err) => console.log(err));
     }
   }
-  //   users.forEach((user) => {
-  //     API.addSubGroupToUser(user._id, subGroupData)
-  //       .then((res) => {
-  //         console.log(res);
-  //         //   loadSubGroup(users);
-  //       })
-  //       .catch((err) => console.log(err));
-  //   });
-  // }
 
-  // loadUser(userState.username)
+  // Adds new subgroup info to outer Group document
   function addSubGroupToGroup(groupID, subGroupData) {
     API.addSubGroupToGroup(groupID, subGroupData)
       .then((res) => {
@@ -110,17 +97,6 @@ function NewGroupModal() {
       })
       .catch((err) => console.log(err));
   }
-
-  //   function loadSubGroup(users) {
-  //     // console.log("before");
-  //     // console.log(username);
-  //     API.getUser(users)
-  //       .then((res) => {
-  //         console.log(res.data);
-  //         setGroupState(res.data);
-  //       })
-  //       .catch((err) => console.log(err));
-  //   }
 
   return (
     <Popup
